fix flow for add and edit activity

+++ backend/src/main/java/com/dalv/bksims/services/social_points_management/ActivityService.java
@@ -73,6 +73,13 @@ public class ActivityService {
             throw new EntityNotFoundException("Organization with name " + organizationRequestName + " not found");
         }
 
+        User owner = userRepo.findById(activityRequest.ownerId()).orElseThrow(
+                () -> new EntityNotFoundException("User with id " + organizationRequestName + " not found"));
+
+        if (owner == null) {
+            throw new EntityNotFoundException("Owner with id " + activityRequest.ownerId() + " not found");
+        }
+
         String organizationName = organization.getName();
 
         // Check validity of dates
@@ -120,10 +127,11 @@ public class ActivityService {
                 .regulationsFileUrl(regulationsFileUrl)
                 .registrationStartDate(activityRequest.registrationStartDate())
                 .registrationEndDate(activityRequest.registrationEndDate())
-                .activityType(activityRequest.activityType())
+                .activityType(activityRequest.activityType().isBlank() ? null : activityRequest.activityType())
                 .status(Status.PENDING.toString())
                 .createdAt(LocalDate.now().toString())
                 .organization(organization)
+                .owner(owner)
                 .build();
 
         activityRepo.save(activity);
@@ -131,7 +139,7 @@ public class ActivityService {
     }
 
     @Transactional
-    public Activity updateActivityInfo(String title, ActivityRequest activityUpdateRequest) {
+    public Activity updateActivityByTitle(String title, ActivityRequest activityUpdateRequest) {
         Activity activity = activityRepo.findOneByTitle(title);
         if (activity == null) {
             throw new EntityNotFoundException("Activity with title " + title + " not found!");
@@ -142,14 +150,15 @@ public class ActivityService {
         activity.setNumberOfParticipants(activityUpdateRequest.numberOfParticipants());
         activity.setCanParticipantsInvite(activityUpdateRequest.canParticipantsInvite());
         activity.setPoints(activityUpdateRequest.points());
+        activity.setActivityType(
+                activityUpdateRequest.activityType().isBlank() ? null : activityUpdateRequest.activityType());
 
         // Check organization name
-        String organizationUpdateRequestName = (activityUpdateRequest.organization() == null) ? activity.getOrganization()
-                .getName() : activityUpdateRequest.organization();
-        Organization organization = organizationRepo.findByName(organizationUpdateRequestName);
+        String organizationRequestName = (activityUpdateRequest.organization() == null) ? "Other School-level Units" : activityUpdateRequest.organization();
+        Organization organization = organizationRepo.findByName(organizationRequestName);
 
         if (organization == null) {
-            throw new EntityNotFoundException("Organization with name " + organizationUpdateRequestName + " not found");
+            throw new EntityNotFoundException("Organization with name " + organizationRequestName + " not found");
         }
 
         String organizationName = organization.getName();
@@ -167,39 +176,53 @@ public class ActivityService {
         activity.setRegistrationStartDate(activityUpdateRequest.registrationStartDate());
         activity.setRegistrationEndDate(activityUpdateRequest.registrationEndDate());
 
-        // Check validity of banner file
+        // Update banner file
         MultipartFile bannerFile = activityUpdateRequest.bannerFile();
-        ActivityValidator.validateBannerFile(bannerFile);
-
-        // Delete old file in S3
+        String currentBannerFileName = activity.getBannerFileName();
         String currentBannerFileUrl = activity.getBannerFileUrl();
-        if (currentBannerFileUrl != null) {
-            s3Service.deleteFileForActivity(currentBannerFileUrl);
+        String bannerFileName = null;
+        String bannerFileUrl = null;
+
+        if (bannerFile == null) {
+            bannerFileName = currentBannerFileName;
+            bannerFileUrl = currentBannerFileUrl;
+        } else {
+            // Check validity of banner file
+            ActivityValidator.validateBannerFile(bannerFile);
+
+            // Delete old file in S3
+            if (currentBannerFileUrl != null) {
+                s3Service.deleteFileForActivity(currentBannerFileUrl);
+            }
+            // Upload new file to S3
+            bannerFileName = s3Service.uploadFileForActivity(bannerFile, organizationName);
+            bannerFileUrl = s3Service.getFileUrl(bannerFileName, organizationName + "/");
         }
-        // Upload new file to S3
-        String bannerFileName = s3Service.uploadFileForActivity(bannerFile, organizationName);
-        String bannerFileUrl = s3Service.getFileUrl(bannerFileName, organizationName + "/");
 
         activity.setBannerFileName(bannerFileName);
         activity.setBannerFileUrl(bannerFileUrl);
 
-        // Check validity of regulations file
-        if (activityUpdateRequest.regulationsFile() != null) {
-            MultipartFile regulationsFile = activityUpdateRequest.regulationsFile();
+        // Update regulations file
+        MultipartFile regulationsFile = activityUpdateRequest.regulationsFile();
+        String currentRegulationsFileUrl = activity.getRegulationsFileUrl();
+        String regulationsFileName = null;
+        String regulationsFileUrl = null;
+
+        if (regulationsFile != null) {
+            // Check validity of regulations file
             ActivityValidator.validateRegulationsFile(regulationsFile);
 
             // Delete old file in S3
-            String currentRegulationsFileUrl = activity.getRegulationsFileUrl();
             if (currentRegulationsFileUrl != null) {
                 s3Service.deleteFileForActivity(currentRegulationsFileUrl);
             }
             // Upload new file to S3
-            String regulationsFileName = s3Service.uploadFileForActivity(regulationsFile, organizationName);
-            String regulationsFileUrl = s3Service.getFileUrl(regulationsFileName, organizationName + "/");
-
-            activity.setRegulationsFileName(regulationsFileName);
-            activity.setRegulationsFileUrl(regulationsFileUrl);
+            regulationsFileName = s3Service.uploadFileForActivity(regulationsFile, organizationName);
+            regulationsFileUrl = s3Service.getFileUrl(regulationsFileName, organizationName + "/");
         }
+
+        activity.setRegulationsFileName(regulationsFileName);
+        activity.setRegulationsFileUrl(regulationsFileUrl);
 
         activityRepo.save(activity);
         return activity;
